Fix copy paste for f64

+++ src/wasm/literal.cpp
@@ -1616,7 +1616,7 @@ Literal Literal::fma(const Literal& left, const Literal& right) const {
       return Literal(fmaf(left.getf32(), right.getf32(), getf32()));
       break;
     case Type::f64:
-      return Literal(fmaf(left.getf64(), right.getf64(), getf64()));
+      return Literal(fma(left.getf64(), right.getf64(), getf64()));
       break;
     default:
       WASM_UNREACHABLE("unexpected type");
@@ -1629,7 +1629,7 @@ Literal Literal::fms(const Literal& left, const Literal& right) const {
       return Literal(fmaf(-left.getf32(), right.getf32(), getf32()));
       break;
     case Type::f64:
-      return Literal(fmaf(-left.getf64(), right.getf64(), getf64()));
+      return Literal(fma(-left.getf64(), right.getf64(), getf64()));
       break;
     default:
       WASM_UNREACHABLE("unexpected type");
